Fixed random server issue
Finally fixed this after forever, turns out that the API was just grabbing random UIDs. When it did, the chance of landing on the same server is higher because servers can be scanned multiple times, adding to the main table. Just made a simple SELECT DISTINCT request to grab a proper list of how many actual unique servers there are and used that as a base on which it can get servers on, then every server has an equal chance of being selected.